Create Sway config.
Currently it's a little buggy, but sway is the future so let's hope for
the best. Reasons to switch to sway:

  1. It's i3 compatible
  2. It runs on wayland
  3. We get to remove all the Xorg shit on my machine (yay)
  4. It's called sway, come on.

Copied i3 config, and then configured it not to run the following stuff
which do not work yet(they're using X for now)

  1. My capslock mapper to Ctrl (NotLikeThis)
  2. Of course redshift does not work yet. :(
  3. Go back to standard i3blocks(temporarily), since I am embedding the
  sway session inside the i3 one(yes you can do this with wayland),
  which bugged my current polybar instances.
  4. Remove clunky keybinds which use x-somethingsomething.
    - Volume control
    - Backlight control

Also figured out how to map capslock to control on wayland, but I should
write such utility since I want to start it as a service.