Add some custom markdowns to gigignore.
Whenever I implement a code review or do some task, I import the whole
thing to a file called TODO.md or when I take notes for a project I work
on, a file called _Notes.md. But when commiting they're getting in the
way of `git add .; git commit` which is very annoying. And yes, I know
there is git add -A(or something), but the dot looks cooler.